Expose CopyGlyphsAcrossFonts
