+++ Dalamud/Interface/GameFonts/GameFontManager.cs
@@ -7,6 +7,7 @@ using System.Text;
 
 using Dalamud.Data;
 using Dalamud.Interface.Internal;
+using Dalamud.Utility;
 using ImGuiNET;
 using Lumina.Data.Files;
 using Serilog;
@@ -111,65 +112,6 @@ namespace Dalamud.Interface.GameFonts
         }
 
         /// <summary>
-        /// Fills missing glyphs in target font from source font, if both are not null.
-        /// </summary>
-        /// <param name="source">Source font.</param>
-        /// <param name="target">Target font.</param>
-        /// <param name="missingOnly">Whether to copy missing glyphs only.</param>
-        /// <param name="rebuildLookupTable">Whether to call target.BuildLookupTable().</param>
-        /// <param name="rangeLow">Low codepoint range to copy.</param>
-        /// <param name="rangeHigh">High codepoing range to copy.</param>
-        public static void CopyGlyphsAcrossFonts(ImFontPtr? source, ImFontPtr? target, bool missingOnly, bool rebuildLookupTable, int rangeLow = 32, int rangeHigh = 0xFFFE)
-        {
-            if (!source.HasValue || !target.HasValue)
-                return;
-
-            var scale = target.Value!.FontSize / source.Value!.FontSize;
-            unsafe
-            {
-                var glyphs = (ImFontGlyphReal*)source.Value!.Glyphs.Data;
-                for (int j = 0, j_ = source.Value!.Glyphs.Size; j < j_; j++)
-                {
-                    var glyph = &glyphs[j];
-                    if (glyph->Codepoint < rangeLow || glyph->Codepoint > rangeHigh)
-                        continue;
-
-                    var prevGlyphPtr = (ImFontGlyphReal*)target.Value!.FindGlyphNoFallback((ushort)glyph->Codepoint).NativePtr;
-                    if ((IntPtr)prevGlyphPtr == IntPtr.Zero)
-                    {
-                        target.Value!.AddGlyph(
-                            target.Value!.ConfigData,
-                            (ushort)glyph->Codepoint,
-                            glyph->X0 * scale,
-                            ((glyph->Y0 - source.Value!.Ascent) * scale) + target.Value!.Ascent,
-                            glyph->X1 * scale,
-                            ((glyph->Y1 - source.Value!.Ascent) * scale) + target.Value!.Ascent,
-                            glyph->U0,
-                            glyph->V0,
-                            glyph->U1,
-                            glyph->V1,
-                            glyph->AdvanceX * scale);
-                    }
-                    else if (!missingOnly)
-                    {
-                        prevGlyphPtr->X0 = glyph->X0 * scale;
-                        prevGlyphPtr->Y0 = ((glyph->Y0 - source.Value!.Ascent) * scale) + target.Value!.Ascent;
-                        prevGlyphPtr->X1 = glyph->X1 * scale;
-                        prevGlyphPtr->Y1 = ((glyph->Y1 - source.Value!.Ascent) * scale) + target.Value!.Ascent;
-                        prevGlyphPtr->U0 = glyph->U0;
-                        prevGlyphPtr->V0 = glyph->V0;
-                        prevGlyphPtr->U1 = glyph->U1;
-                        prevGlyphPtr->V1 = glyph->V1;
-                        prevGlyphPtr->AdvanceX = glyph->AdvanceX * scale;
-                    }
-                }
-            }
-
-            if (rebuildLookupTable)
-                target.Value!.BuildLookupTable();
-        }
-
-        /// <summary>
         /// Unscales fonts after they have been rendered onto atlas.
         /// </summary>
         /// <param name="fontPtr">Font to unscale.</param>
@@ -191,7 +133,7 @@ namespace Dalamud.Interface.GameFonts
                 font->Descent /= fontScale;
                 if (font->ConfigData != null)
                     font->ConfigData->SizePixels /= fontScale;
-                var glyphs = (ImFontGlyphReal*)font->Glyphs.Data;
+                var glyphs = (Util.ImFontGlyphReal*)font->Glyphs.Data;
                 for (int i = 0, i_ = font->Glyphs.Size; i < i_; i++)
                 {
                     var glyph = &glyphs[i];
@@ -260,7 +202,7 @@ namespace Dalamud.Interface.GameFonts
         /// <param name="rebuildLookupTable">Whether to call target.BuildLookupTable().</param>
         public void CopyGlyphsAcrossFonts(ImFontPtr? source, GameFontStyle target, bool missingOnly, bool rebuildLookupTable)
         {
-            GameFontManager.CopyGlyphsAcrossFonts(source, this.fonts[target], missingOnly, rebuildLookupTable);
+            Util.CopyGlyphsAcrossFonts(source, this.fonts[target], missingOnly, rebuildLookupTable);
         }
 
         /// <summary>
@@ -272,7 +214,7 @@ namespace Dalamud.Interface.GameFonts
         /// <param name="rebuildLookupTable">Whether to call target.BuildLookupTable().</param>
         public void CopyGlyphsAcrossFonts(GameFontStyle source, ImFontPtr? target, bool missingOnly, bool rebuildLookupTable)
         {
-            GameFontManager.CopyGlyphsAcrossFonts(this.fonts[source], target, missingOnly, rebuildLookupTable);
+            Util.CopyGlyphsAcrossFonts(this.fonts[source], target, missingOnly, rebuildLookupTable);
         }
 
         /// <summary>
@@ -284,7 +226,7 @@ namespace Dalamud.Interface.GameFonts
         /// <param name="rebuildLookupTable">Whether to call target.BuildLookupTable().</param>
         public void CopyGlyphsAcrossFonts(GameFontStyle source, GameFontStyle target, bool missingOnly, bool rebuildLookupTable)
         {
-            GameFontManager.CopyGlyphsAcrossFonts(this.fonts[source], this.fonts[target], missingOnly, rebuildLookupTable);
+            Util.CopyGlyphsAcrossFonts(this.fonts[source], this.fonts[target], missingOnly, rebuildLookupTable);
         }
 
         /// <summary>
@@ -449,7 +391,7 @@ namespace Dalamud.Interface.GameFonts
                     }
                 }
 
-                CopyGlyphsAcrossFonts(InterfaceManager.DefaultFont, font, true, false);
+                Util.CopyGlyphsAcrossFonts(InterfaceManager.DefaultFont, font, true, false);
                 UnscaleFont(font, 1 / scale, false);
                 font.BuildLookupTable();
             }
@@ -468,38 +410,6 @@ namespace Dalamud.Interface.GameFonts
 
                 if ((this.fontUseCounter[style] -= 1) == 0)
                     this.fontUseCounter.Remove(style);
-            }
-        }
-
-        private struct ImFontGlyphReal
-        {
-            public uint ColoredVisibleCodepoint;
-            public float AdvanceX;
-            public float X0;
-            public float Y0;
-            public float X1;
-            public float Y1;
-            public float U0;
-            public float V0;
-            public float U1;
-            public float V1;
-
-            public bool Colored
-            {
-                get => ((this.ColoredVisibleCodepoint >> 0) & 1) != 0;
-                set => this.ColoredVisibleCodepoint = (this.ColoredVisibleCodepoint & 0xFFFFFFFEu) | (value ? 1u : 0u);
-            }
-
-            public bool Visible
-            {
-                get => ((this.ColoredVisibleCodepoint >> 1) & 1) != 0;
-                set => this.ColoredVisibleCodepoint = (this.ColoredVisibleCodepoint & 0xFFFFFFFDu) | (value ? 2u : 0u);
-            }
-
-            public int Codepoint
-            {
-                get => (int)(this.ColoredVisibleCodepoint >> 2);
-                set => this.ColoredVisibleCodepoint = (this.ColoredVisibleCodepoint & 3u) | ((uint)this.Codepoint << 2);
             }
         }
     }
